requirements.txt
streamlit>=1.44.0
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
